Refactor, moved Cell class inside Board
